template/orion_mlops_template/{{cookiecutter.project_slug}}/src/dev-requirements.txt
jupyter~=1.0
jupyter_client>=5.1, <7.0
jupyterlab~=3.0
kedro-viz==4.3.1